Modyfikacja po 2 lekcji

+++ include/wpisy_braznowe.h
@@ -3,13 +3,13 @@
 class WpisyBranzowe {
 
 public:
-	WpisyBranzowe(std::string, std::string, std::string int, int);
+	WpisyBranzowe(std::string, std::string, std::string, float, float);
 
 	std::string kategoria;
 	std::string nazwa
 	std::string data;
-	int cenaAkcji;
-	int zmiana;
+	float cenaAkcji;
+	float zmiana;
 	
 	
 };
